fixed some errors
there was some errors In BFS search when it check for the next node, as it was trying to access the currentNode of the current_Node.parent of the initial one which wasn't have any  parent. how ever there still some bugs in the AIPlayerController that it's not working as it should, but I am still working on.